rename bindings.rs -> glue.rs to hint at how we got it
