+++ src/rbt.rs
@@ -1,4 +1,4 @@
-use crate::bindings;
+use crate::glue;
 use itertools::Itertools;
 use roc_std::{RocList, RocStr};
 use serde::de::Visitor;
@@ -12,8 +12,8 @@ pub struct Rbt {
     pub default: Job,
 }
 
-impl From<bindings::Rbt> for Rbt {
-    fn from(rbt: bindings::Rbt) -> Self {
+impl From<glue::Rbt> for Rbt {
+    fn from(rbt: glue::Rbt) -> Self {
         let unwrapped = rbt.f0;
 
         Rbt {
@@ -25,7 +25,7 @@ impl From<bindings::Rbt> for Rbt {
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Job {
     pub command: Command,
-    pub inputs: HashMap<String, Job>, // TODO: when this can be generated by bindings it will need to be a RocList
+    pub inputs: HashMap<String, Job>, // TODO: when this can be generated by glue it will need to be a RocList
     #[serde(
         serialize_with = "serialize_roc_list_of_roc_str",
         deserialize_with = "deserialize_roc_list_of_roc_str"
@@ -38,8 +38,8 @@ pub struct Job {
     pub outputs: RocList<RocStr>,
 }
 
-impl From<bindings::Job> for Job {
-    fn from(job: bindings::Job) -> Self {
+impl From<glue::Job> for Job {
+    fn from(job: glue::Job) -> Self {
         // let unwrapped = job.into_Job();
         let unwrapped = job.f0;
 
@@ -82,8 +82,8 @@ pub struct Command {
     pub args: RocList<RocStr>,
 }
 
-impl From<bindings::Command> for Command {
-    fn from(command: bindings::Command) -> Self {
+impl From<glue::Command> for Command {
+    fn from(command: glue::Command) -> Self {
         // let unwrapped = command.into_Job();
         let unwrapped = command.f0;
 
@@ -105,8 +105,8 @@ pub enum Tool {
     },
 }
 
-impl From<bindings::Tool> for Tool {
-    fn from(tool: bindings::Tool) -> Self {
+impl From<glue::Tool> for Tool {
+    fn from(tool: glue::Tool) -> Self {
         Self::SystemTool { name: tool.f0 }
     }
 }
